WIP: RP initiated logout
