+++ packages/core/src/index.ts
@@ -92,6 +92,11 @@ export { InvalidResponseError } from "./errors/InvalidResponseError";
 export { OidcProviderError } from "./errors/OidcProviderError";
 
 export { default as GeneralLogoutHandler } from "./logout/GeneralLogoutHandler";
+export {
+  clear,
+  getUnauthenticatedSession,
+  SessionInfoManagerBase
+} from "./sessionInfo/SessionInfoManager"
 
 export {
   createDpopHeader,
