+++ packages/core/src/logout/GeneralLogoutHandler.spec.ts
@@ -20,7 +20,7 @@
 //
 
 import { it, describe, expect } from "@jest/globals";
-import { mockStorageUtility } from "@inrupt/solid-client-authn-core";
+import { mockStorageUtility } from "../mocks";
 import LogoutHandler from "./GeneralLogoutHandler";
 import { mockSessionInfoManager } from "../sessionInfo/__mocks__/SessionInfoManager";
