+++ packages/node/src/sessionInfo/SessionInfoManager.ts
@@ -23,93 +23,20 @@
  * @hidden
  * @packageDocumentation
  */
-
 import {
   ISessionInfo,
   ISessionInternalInfo,
   ISessionInfoManager,
-  ISessionInfoManagerOptions,
-  IStorageUtility,
+  SessionInfoManagerBase,
+  clear
 } from "@inrupt/solid-client-authn-core";
-import { v4 } from "uuid";
-// eslint-disable-next-line no-shadow
-import { fetch } from "@inrupt/universal-fetch";
 import { KEY_REGISTERED_SESSIONS } from "../constant";
-
-export function getUnauthenticatedSession(): ISessionInfo & {
-  fetch: typeof fetch;
-} {
-  return {
-    isLoggedIn: false,
-    sessionId: v4(),
-    fetch,
-  };
-}
-
-/**
- * @param sessionId
- * @param storage
- * @hidden
- */
-export async function clear(
-  sessionId: string,
-  storage: IStorageUtility
-): Promise<void> {
-  await Promise.all([
-    storage.deleteAllUserData(sessionId, { secure: false }),
-    storage.deleteAllUserData(sessionId, { secure: true }),
-    // FIXME: This is needed until the DPoP key is stored safely
-    storage.delete("clientKey", { secure: false }),
-  ]);
-  // TODO: Clear OIDC storage if need be.
-  // await clearOidcPersistentStorage();
-}
+export { getUnauthenticatedSession, clear } from "@inrupt/solid-client-authn-core"
 
 /**
  * @hidden
  */
-export class SessionInfoManager implements ISessionInfoManager {
-  constructor(private storageUtility: IStorageUtility) {}
-
-  // eslint-disable-next-line class-methods-use-this
-  update(
-    _sessionId: string,
-    _options: ISessionInfoManagerOptions
-  ): Promise<void> {
-    // const localUserId: string = options.localUserId || this.uuidGenerator.v4();
-    // if (options.loggedIn) {
-    //   return {
-    //     sessionId,
-    //     loggedIn: true,
-    //     webId: options.webId as string,
-    //     neededAction: options.neededAction || { actionType: "inaction" },
-    //     state: options.state,
-    //     logout: async (): Promise<void> => {
-    //       // TODO: handle if webid isn't here
-    //       return this.logoutHandler.handle(localUserId);
-    //     },
-    //     fetch: (url: RequestInfo, init?: RequestInit): Promise<Response> => {
-    //       // TODO: handle if webid isn't here
-    //       return this.authenticatedFetcher.handle(
-    //         {
-    //           localUserId,
-    //           type: "dpop"
-    //         },
-    //         url,
-    //         init
-    //       );
-    //     }
-    //   };
-    // } else {
-    //   return {
-    //     localUserId,
-    //     loggedIn: false,
-    //     neededAction: options.neededAction || { actionType: "inaction" }
-    //   };
-    // }
-    throw new Error("Not Implemented");
-  }
-
+export class SessionInfoManager extends SessionInfoManagerBase implements ISessionInfoManager {
   async get(
     sessionId: string
   ): Promise<(ISessionInfo & ISessionInternalInfo) | undefined> {
@@ -135,11 +62,6 @@ export class SessionInfoManager implements ISessionInfoManager {
     }
 
     return undefined;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  async getAll(): Promise<(ISessionInfo & ISessionInternalInfo)[]> {
-    throw new Error("Not implemented");
   }
 
   /**
